Dicts and Functions
To facilitate changes from 7 to 8, I've introduced some new functions and rearranged the dictionaries, as alluded to in previous commits. I've also partially split out the functionality of 7b so that the parameter combinations are handled by themselves in one place. A future file will be 8a2 (or similar variant) where we go back into the spatial domain. This should be easier with these changes as just copying 8a, making edits to the copy, and switching the file loaded into 8b.