Create TimeSpaceAndTimeSeries-8g-Plots.R
Actually developed this file a week and a half ago in a rush. Now going back through it.

Inside is a function that uses the dictionaries as global variables (hence the sourcing) to convert the file parameters to human-readables.

It then loads the diversity files, unifies the time format (which should be fixed at the root anyways, and is just catching any mistakes made with the previous version), and tries to flatten the files to a single data.frame.

Doing so requires reformatting to a common format, then applying human readable versions of the columns. Note that some false NAs are present due to the way the diversity values are being calculated, specifically in the richness case, so those are corrected.

The unfinished/not completely tested bits are then about the averaging to reduce visual clutter (lines 177 - 202). The plotting function that we end with appears functional at the moment, but some of the results that I've been observing do look anomalous (see notes).