Why's the Diversity So Low? Well...
TL;DR: Line 46 in 8b.

The original publication and system was running with around 5 species or so on average, but we've been recently getting around 2. Why is the diversity now so low? To check, I've been re-running the old system (retryC311) with different variations of the parameters (8a2 with debugonce-based injections) and visualising them (compareC311). The conclusion is that the main difference is an accidental conflict in how I was trying to re-decide to implement the colonisation density, so that species were coming in with much more abundance than the system could hope to sustain. My new vision had been to do a multiplication automatically, rather than requiring the user to do so, but I had failed to fully implement this.

As a part of the process, I also implemented initial conditions as a dictionary of interest and made some robustness edits and corrections to the code.