Create CompareEliminationThresholds-Script.R
Trying to determine how the elimination threshold varies the presence or absence of species. If I've set the threshold and propagule size well, then there shouldn't be much effect. If there is, that suggests that we need to have a good think here about what the right level is, especially given we want to increase the number of species.

For what it's worth, the number of species that we're getting isn't really far from LM1996. Images seem to suggest that it's not so much the number of species that's changing, but the species identities.