Logistic Intervention Prototype
Added handling so that if the previous rprime is a function, we extract the reproduction rates from it. This is then used in the switches as previously, but then we need to rewrap it inside the logistic functional form before continuing as normal.

(So now we have a function (MNA) that calls the ode solver, which is called by the set-up wrapper 9d, which is called by another script 9e, and the solver now uses a function (PerCapitaDynamics) which calls a function (rprimef) of a function (rprimeMax = rprimef as switches). I think I've got all of that?)