Update TimeSpaceAndTimeSeries-9h-CombinedPlots.R
Added filtering for timings so that we can be comparing consistent amounts of simulation to each other. Added a grey rectangle recognizing the region.

Made the in and out distinctions clearer in the data itself.

Added in the present and failed arrivals via ColExt.

Likely to remove plot1 in favor of 2.