Suggestion of some long runs
JWP would like me to check for convergence if we run the simulations for long enough. I'm running 3 runs with the same set-up but different starting land-uses. Then I'll use interventions to (0) (from (0), (0.5), and (1)) and check to see if, running it for 190 instead of the usual 10 after the 10 setup (line 41), it eventually reaches the same distribution regardless of starting state. (Original runs show a decrease of around 5-6 species for (0.5) and (1) respectively.)