Update TimeSpaceAndTimeSeries-9h-CombinedPlots.R
Playing around with updating the basic overview plot. Currently I'm happy with panels 1 and 2, but I'm unsure about panel 3 (and if it should be a single panel, or multipart, such as scatter plots of abundance (consumer vs basal), violin plots of temporal beta diversity, and either size or consumer/basal ratio.