Update EliminationAndNeutralEvents.R
I suspect line 43 is the answer to the previous commit. In the case where the system changes land-use types, I've set a few of my PCD functions to be time sensitive, but this line defaults to whatever happened at time 0. This doesn't affect the previous publication or any of my base-case results, but it does affect the perceived ability to invade after land-use changes...