Simple Statistical Testing
As a first attempt, this is fine, but I'm not happy with it yet. This is the problem that the realignment in the diversity metrics is meant to be a partial solution of, in that there might be a bias in the simple gls models recommended online due to the slight offset. Equally, I worry that the models may be insufficient with the simple AR1 structure, so I'm also hoping to address that.