Update TimeSpaceAndTimeSeries-9h-CombinedPlots.R
The basal consumer split might be excessive here, but might help present a more nuanced/balanced view. The annoying bit is that it necessarily means a bit of artifice as to what no consumers means for an ecosystem compared to an ecosystem with consumers, as well as how to combine the two separate distances (choosing L1).

Running overnight to see if we get nicer results from any of the algorithms.