Create TimeSpaceAndTimeSeries-9h2-CombinedPlots.R
Setting up the base for recreating some of the plots. I am getting to a point where I'm in desperate need of a refactor admittedly (moving functions and communal resources around). As we're approaching publication, the variant experiments as a whole will need to be moved into the main package, which is likely when I'll be sorting through things here and there.