Addition of minTimePlus argument to help F4
Collecting thoughts on how I think the Figure 4 should work. Unfortunately, given the conversations I've had, we probably need to rerun the diversities (which might mean I can fix one other nibbling thing that has been bothering me anyways, so that's not necessarily as large of a loss as it would otherwise be).