Refactor Affinity Dictionary for better seeding
Practically speaking, this is just a reset of the seeds combined with a shift from one dictionary handling all of the affinities to two, one for species and one for patches.

Implementation wise, this requires re-evaluating pretty much every instance of anything touching the affinities to make sure it is calling for the right aspect.