Update TimeSpaceAndTimeSeries-9h2-CombinedPlots.R
(Lines 26 - 32, small change in spacing.)

Unfortunately, putting graphs on top of kernel density estimates appears to be a bit futile in terms of getting the two packages to fully talk to each other correctly regarding facets (nodes and edges repeatedly ending up on different panels from each other and the kdes they are supposed to be represented by). The KDEs by themselves are probably sufficient though.