Update TimeSpaceAndTimeSeries-9h2-CombinedPlots.R
I've been asked to change the color scheme and to move the 2a inset times. I'm calling it quits on getting the coordinate scheme perfect (i.e., consistent and visually nice). I've also been asked to make the logic of the plots visually clearer and to try to generalise Figure 5.