Update TimeSpaceAndTimeSeries-10j-Supplement10.R
I think I've found an interesting supplemental figure. It captures how most of the interactions at the 100% 0 in (1) situation is at the smallest sizes, while it is evenly spread through the 100% 0 in (0.5) system. The 100% 0 in (0) system shows some nonlinearities at the highest levels that might be indicative of a difference in the consumer-consumer relationships.