Double checked Carrying Capacity & Biomass
Carrying capacity is almost essentially the same in terms of no intervention richness. There seems to be some effect in terms of raw richness (sure) and a bit more of an effect in the case 0's in (0.25) where the carrying capacity seems to prevent the basals and consumers from getting out of hand.

Biomass and abundance follow the roughly nonlinear and cone shaped trends that we saw earlier. There are some things to double check, but at first glance it seems fine.