Expanded Supplement 2 to include abundance
This meant combining the previous version with next previous version to get combined figured. I also added a switch to navigate which interventions to consider. Note that for each initial state, we get essentially the same final state.

(These still work with the carrying capacity version, but those don't cover all cases, so the initial == 1 case is the one that makes the most sense.)

And fixed file names.